还差应用端的缓存共享没做

+++ Dao/CacheInit.go
@@ -1,6 +1,7 @@
 package Dao
 
 import (
+	"../util"
 	"./Cache"
 	"./Database"
 )
@@ -9,13 +10,22 @@ import (
 //
 //初始化过程中有任何出错都返回error，顺利完成就返回nil
 func CacheInit() error {
-	_, _ = Cache.Flushall()
-	rows, err := Database.GetALL(Cache.Conf.InitialCache)
-	if err != nil {
-		return err
-	}
-	for ID, PASS := range rows {
-		_, _ = Cache.SetUser(ID, PASS)
-	}
+	if Cache.Conf.MeetAtStart {
+		_, _ = Cache.Flushall()
+		rows, err := Database.GetALL(Cache.Conf.InitialCache)
+		util.LogE(err)
+		for ID, PASS := range rows {
+			_, err = Cache.SetUser(ID, PASS)
+			util.LogE(err)
+		}
+	} /*TODO:集群接口暂不支持
+	if Cache.Conf.MeetAtStart {
+		db, err := Cache.RedisConnect()
+		util.LogE(err)
+		if db != nil {
+			_, err = db.Do("CLUSTER", "MEET", Cache.Conf.MeetAddress, Cache.Conf.MeetPort)
+			util.LogE(err)
+		}
+	}*/
 	return nil
 }
